refactor(editor): extract BufferLifecycle for buffer registration and file opening
Move buffer registration, session restore, swap recovery, file opening,
tab lookup, notification helpers, and tool picker refresh into
MingaEditor.Handlers.BufferLifecycle. The editor keeps defdelegate stubs
for all 9 public functions to preserve backward compatibility.

Editor reduced from 1,565 to 1,255 lines (-310).

Part of #1433 (carve mega-modules into focused submodules).
